requirements.txt
fastapi~=0.68.1
pydantic~=2.5.2
numpy~=1.24.3
openai~=1.55.3
python-dotenv~=1.0.0
pypdf~=3.17.1
pytest~=7.4.3
uvicorn~=0.15.0
sentence-transformers~=3.3.1
tqdm~=4.67.1